refactor(desktop): decompose main.rs into sidecar, file_save, and alloc_probe modules
main.rs had grown to ~4100 lines spanning the sidecar lifecycle, the streaming
file-save subsystem, and a test-only peak allocator alongside the shell core.
Extract three cohesive, nearly self-contained clusters into their own modules so
each concern lives in one place and main.rs shrinks to the shell core + IPC
command surface (~3000 lines).

- sidecar.rs: sidecar process bootstrap/connect/handshake/shutdown -- the layer
  above `sidecar_ipc` (transport) and `frame` (codec). Owns SidecarBootstrap,
  bootstrap_sidecar / bootstrap_dev_sidecar / spawn_stdio_sidecar,
  try_connect_dev_sidecar, the unix connect_and_handshake_dev_sidecar twin,
  request_sidecar_shutdown, fetch_sidecar_info, write_sidecar_metadata, and
  hash_sidecar_binary. The Windows handshake twin stays in windows_impl; this
  module re-exports the unix twin and main.rs re-exports the Windows one so both
  are reachable as crate-root siblings.
- file_save.rs: the streaming file-save subsystem (file_save_open[_dialog] →
  write* → commit | abort), its registry/handle structs, the tmp/defuse/partial
  helpers, and the orphan-sweep + idle-handle GC. Self-contained except for
  `decode_b64` and `recover`, which it imports from main.rs.
- alloc_probe.rs: the test-only peak-tracking allocator, lifted out of an inline
  `mod alloc_probe` so the allocator plumbing has its own file.

Visibility bumps: SidecarMetadata, the SIDECAR_* consts/env-vars, and
decode_b64 become pub(crate); the save subsystem's registry/handle surface and
the sidecar bootstrap entry become pub(crate) so main.rs (and its tests) can
reach them. The runtime sidecar handle (SidecarProcess + reader/writer threads)
stays in main.rs with DesktopShell -- it is a different concern (the live
shell's handle on a connected sidecar, not bootstrapping one).

Verified: cargo +1.97.1 clippy --all-targets -- -D warnings clean; 55 Rust tests
pass; cargo fmt clean; Windows cross-compile clean (only the pre-existing
embed_plist Tauri artifact remains) and the windows_impl/sidecar modules
type-check clean on the windows target. Windows CI is the final gate for the
Windows-only Rust paths.
